introducing boolean parameter in application.conf to disable Solr

+++ app/solr/Delete.java
@@ -8,6 +8,9 @@ import models.User;
 
 import org.apache.solr.client.solrj.SolrServerException;
 
+import play.Logger;
+import play.Play;
+
 public class Delete {
 
 	public static void deleteAll(){
@@ -39,13 +42,19 @@ public class Delete {
 	}
 	
 	public static void deleteByQuery(String Query){
+		Boolean enabled = Boolean.valueOf(Play.configuration.getProperty("milk.solr.enabled", "false"));
+		if ( !enabled ) {
+			Logger.info("Tried to delete document from Solr, while it's disabled");
+			return;
+		}
+		
 		try {
 			Server.getServer().deleteByQuery(Query);
 			Server.getServer().commit();
 		} catch (SolrServerException e) {
-			// TODO
+			Logger.warn("Exception while deleting documents from Solr", e);
 		} catch (IOException e) {
-			// TODO
+			Logger.warn("Exception while deleting documents from Solr", e);
 		}
 	}
